requirements.txt
pandas==2.3.3
yfinance==0.2.66
scikit-learn==1.7.2
python-dotenv>=1.0.0
torch>=2.0            
pytorch-lightning>=2.0
torchmetrics>=0.11
mlflow>=2.0
fastapi==0.119.1
uvicorn==0.38.0
PyJWT==2.10.1
python-multipart==0.0.20
prometheus_client==0.23.1
psutil==7.1.3